10/26 3:01 pm
Added more functions and added all of the algorithm from the handout. Still working out bugs because for some reason it is running many more times than it should. I think the r[i] thing might not be working right (i.e. it isn't updating like it should in the functions).